backend/requirements.txt
django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-mptt==0.15.0
django-cors-headers==4.4.0
redis==5.0.7
django-redis==6.2.0
celery==5.4.0
requests==2.32.3
python-dotenv==1.0.1
